feat(kiloclaw): add org CLI run page
